+++ libraries/Monobjc.Foundation/Foundation_S/NSSize.cs
@@ -23,6 +23,7 @@
 using System;
 using System.Globalization;
 using System.Runtime.InteropServices;
+using Monobjc.ApplicationServices;
 
 namespace Monobjc.Foundation
 {
@@ -50,7 +51,7 @@ namespace Monobjc.Foundation
         /// </summary>
         /// <param name="width">The width.</param>
         /// <param name="height">The height.</param>
-        public NSSize(float width, float height)
+		public NSSize(CGFloat width, CGFloat height)
         {
             this.width = width;
             this.height = height;
